apisnapshot set in history
